+++ lib/ex_worker/server.ex
@@ -1,10 +1,16 @@
 defmodule ExWorker.Server do
   use GenServer
 
+  @message_server_pid ExWorker.MessageServer.start
+
   # GenServer API
 
   # init server
-  def init(state), do: {:ok, state}
+  def init(state) do
+    schedule_work()
+
+    {:ok, state}
+  end
 
   # take first message from state
   def handle_call(:take_message, _, [value | state]), do: {:reply, value, state}
@@ -19,6 +25,26 @@ defmodule ExWorker.Server do
   defp place_value_to_end(state, value) do
     [value | Enum.reverse(state)] |> Enum.reverse()
   end
+
+  # receive handle message result from message server
+  def handle_info({:send_message_result, _}, state) do
+    {:noreply, state}
+  end
+
+  # scheduler work
+  def handle_info(:work, state) do
+    IO.puts "Execute scheduler"
+
+    caller = self()
+    send(@message_server_pid, {:send_message, caller, "SOMETHING"})
+
+    schedule_work()
+
+    {:noreply, state}
+  end
+
+  # Run schedule n 5 seconds
+  defp schedule_work, do: Process.send_after(self(), :work, 5 * 1000)
 
   # Client API